All common parse code was moved to lgparse.py.
